raktsetu_P1/functions/requirements.txt
firebase-functions>=0.1.0
firebase-admin>=6.2.0
google-cloud-firestore
google-cloud-aiplatform
google-cloud-bigquery
requests